+++ changelog.md
@@ -10,6 +10,8 @@
 * [Fixed] mode and order in `resize_to_shape` function
 * [Fixed] respect new functionality in skimage.transform
 * [Added] New arguments in `resize_to_shape` function
+* [Added] New tests for seeds
+* [Added] On resize check if all seeds are preserved
 
 ## 1.26.0